+++ libs/designer/src/lib/common/constants.ts
@@ -454,6 +454,10 @@ export default {
     SWAGGER_ENDPOINT: 'SWAGGER_ENDPOINT',
     TESTING: 'TESTING',
   },
+  ERRORS_PANEL_TAB_NAMES: {
+    ERRORS: 'ERRORS',
+    WARNINGS: 'WARNINGS',
+  },
   PROFILE_KEY_TYPE: {
     DESIGNER: 'designer:',
     INITIALIZE_NODES: 'designer:initializeNodes',
@@ -483,6 +487,24 @@ export default {
   DEFAULT_RECURRENCE: {
     interval: 3,
     frequency: 'Minute',
+  },
+  RECURRENCE_OPTIONS: {
+    FREE: {
+      interval: 1,
+      frequency: 'Hour',
+    },
+    STANDARD: {
+      interval: 1,
+      frequency: 'Minute',
+    },
+    PREMIUM: {
+      interval: 15,
+      frequency: 'Second',
+    },
+    CONSUMPTION: {
+      interval: 3,
+      frequency: 'Minute',
+    },
   },
   RECURRENCE_FREQUENCY_VALUES: ['Month', 'Week', 'Day', 'Hour', 'Minute', 'Second'],
   RECURRENCE_TITLE_JOIN_SEPARATOR: ',',