+++ libs/designer/src/lib/core/actions/bjsworkflow/add.ts
@@ -4,7 +4,7 @@ import { getConnectionsForConnector, getConnectorWithSwagger } from '../../queri
 import { getOperationManifest } from '../../queries/operation';
 import { initEmptyConnectionMap } from '../../state/connection/connectionSlice';
 import type { NodeData, NodeOperation, OperationMetadataState } from '../../state/operation/operationMetadataSlice';
-import { initializeNodes, initializeOperationInfo } from '../../state/operation/operationMetadataSlice';
+import { initializeNodes, initializeOperationInfo, updateNodeSettings } from '../../state/operation/operationMetadataSlice';
 import type { RelationshipIds } from '../../state/panel/panelInterfaces';
 import { changePanelNode, openPanel, setIsPanelLoading } from '../../state/panel/panelSlice';
 import { addResultSchema } from '../../state/staticresultschema/staticresultsSlice';
@@ -134,7 +134,8 @@ export const initializeOperationDetails = async (
       manifest,
       /* swagger */ undefined,
       /* operation */ undefined,
-      state.workflow.workflowKind
+      state.workflow.workflowKind,
+      state.designerOptions.hostOptions.forceEnableSplitOn
     );
 
     // We should update the outputs when splitOn is enabled.
@@ -188,7 +189,8 @@ export const initializeOperationDetails = async (
       /* manifest */ undefined,
       parsedSwagger,
       /* operation */ undefined,
-      state.workflow.workflowKind
+      state.workflow.workflowKind,
+      state.designerOptions.hostOptions.forceEnableSplitOn
     );
 
     // We should update the outputs when splitOn is enabled.
@@ -243,7 +245,9 @@ export const initializeOperationDetails = async (
   const triggerNodeManifest = await getTriggerNodeManifest(state.workflow, state.operations);
 
   if (triggerNodeManifest) {
-    updateInvokerSettings(isTrigger, triggerNodeManifest, nodeId, initData.settings as Settings, dispatch);
+    updateInvokerSettings(isTrigger, triggerNodeManifest, initData.settings as Settings, (invokerSettings: Settings) =>
+      dispatch(updateNodeSettings({ id: nodeId, settings: invokerSettings }))
+    );
   }
 
   updateAllUpstreamNodes(getState() as RootState, dispatch);