+++ libs/designer/src/lib/core/actions/bjsworkflow/initialize.ts
@@ -296,7 +296,8 @@ export const updateOutputsAndTokens = async (
   inputs: NodeInputs,
   settings: Settings,
   shouldProcessSettings = false,
-  workflowKind?: WorkflowKind
+  workflowKind?: WorkflowKind,
+  forceEnableSplitOn?: boolean
 ): Promise<void> => {
   const { type, kind, connectorId } = operationInfo;
   const supportsManifest = OperationManifestService().isSupported(type, kind);
@@ -332,7 +333,7 @@ export const updateOutputsAndTokens = async (
   dispatch(updateTokens({ id: nodeId, tokens }));
 
   // NOTE: Split On setting changes as outputs of trigger changes, so we will be recalculating such settings in this block for triggers.
-  if (shouldProcessSettings && isTrigger && workflowKind !== WorkflowKind.STATELESS) {
+  if (shouldProcessSettings && isTrigger && (workflowKind !== WorkflowKind.STATELESS || forceEnableSplitOn)) {
     const isSplitOnSupported = getSplitOnOptions(nodeOutputs, supportsManifest).length > 0;
     if (settings.splitOn?.isSupported !== isSplitOnSupported) {
       dispatch(updateNodeSettings({ id: nodeId, settings: { splitOn: { ...settings.splitOn, isSupported: isSplitOnSupported } } }));
@@ -525,20 +526,19 @@ const getSwaggerFromService = async (serviceDetails: CustomSwaggerServiceDetails
 
 export const updateInvokerSettings = (
   isTrigger: boolean,
-  tiggerNodeManifest: OperationManifest | undefined,
-  nodeId: string,
+  triggerNodeManifest: OperationManifest | undefined,
   settings: Settings,
-  dispatch: Dispatch,
+  updateNodeSettingsCallback: (invokerSettings: Settings) => void,
   references?: ConnectionReferences
 ): void => {
-  if (!isTrigger && tiggerNodeManifest?.properties?.settings?.invokerConnection) {
-    dispatch(updateNodeSettings({ id: nodeId, settings: { invokerConnection: { ...settings.invokerConnection, isSupported: true } } }));
+  if (!isTrigger && triggerNodeManifest?.properties?.settings?.invokerConnection) {
+    updateNodeSettingsCallback({ invokerConnection: { ...settings.invokerConnection, isSupported: true } });
   }
   if (references) {
     Object.keys(references).forEach((key) => {
       const impersonationSource = references[key].impersonation?.source;
       if (impersonationSource === ImpersonationSource.Invoker) {
-        dispatch(updateNodeSettings({ id: nodeId, settings: { invokerConnection: { isSupported: true, value: { enabled: true } } } }));
+        updateNodeSettingsCallback({ invokerConnection: { isSupported: true, value: { enabled: true } } });
       }
     });
   }