+++ libs/designer/src/lib/core/actions/bjsworkflow/settings.ts
@@ -119,7 +119,8 @@ export const getOperationSettings = (
   manifest?: OperationManifest,
   swagger?: SwaggerParser,
   operation?: LogicAppsV2.OperationDefinition,
-  workflowKind?: WorkflowKind
+  workflowKind?: WorkflowKind,
+  forceEnableSplitOn?: boolean
 ): Settings => {
   const { operationId, type: nodeType } = operationInfo;
   return {
@@ -148,8 +149,8 @@ export const getOperationSettings = (
       value: getDisableAutomaticDecompression(isTrigger, nodeType, manifest, operation),
     },
     splitOn: {
-      isSupported: isSplitOnSupported(isTrigger, nodeOutputs, manifest, swagger, operationId, operation, workflowKind),
-      value: getSplitOn(manifest, swagger, operationId, operation, workflowKind),
+      isSupported: isSplitOnSupported(isTrigger, nodeOutputs, manifest, swagger, operationId, operation, workflowKind, forceEnableSplitOn),
+      value: getSplitOn(manifest, swagger, operationId, operation, workflowKind, forceEnableSplitOn),
     },
     retryPolicy: {
       isSupported: isRetryPolicySupported(isTrigger, operationInfo.type, manifest),
@@ -165,7 +166,7 @@ export const getOperationSettings = (
       value: getSuppressWorkflowHeaders(isTrigger, nodeType, manifest, operation),
     },
     suppressWorkflowHeadersOnResponse: {
-      isSupported: isSuppressWorklowHeadersOnResponseSupported(isTrigger, manifest),
+      isSupported: isSuppressWorkflowHeadersOnResponseSupported(isTrigger, manifest),
       value: getSuppressWorkflowHeadersOnResponse(operation),
     },
     concurrency: {
@@ -343,7 +344,7 @@ const getSuppressWorkflowHeadersOnResponse = (definition?: LogicAppsV2.Operation
   return isOperationOptionSet(Constants.SETTINGS.OPERATION_OPTIONS.SUPPRESS_WORKFLOW_HEADERS_ON_RESPONSE, operationOptions);
 };
 
-const isSuppressWorklowHeadersOnResponseSupported = (isTrigger: boolean, manifest?: OperationManifest): boolean => {
+const isSuppressWorkflowHeadersOnResponseSupported = (isTrigger: boolean, manifest?: OperationManifest): boolean => {
   if (manifest) {
     const operationOptionsSetting = getOperationSettingFromManifest(manifest, 'operationOptions') as
       | OperationManifestSetting<OperationOptions[]>
@@ -516,9 +517,12 @@ const getSplitOn = (
   swagger?: SwaggerParser,
   operationId?: string,
   definition?: LogicAppsV2.OperationDefinition,
-  workflowKind?: WorkflowKind
+  workflowKind?: WorkflowKind,
+  forceEnableSplitOn?: boolean
 ): SimpleSetting<string> => {
-  if (workflowKind === WorkflowKind.STATELESS) return { enabled: false };
+  if (workflowKind === WorkflowKind.STATELESS && !forceEnableSplitOn) {
+    return { enabled: false };
+  }
   const splitOnValue = getSplitOnValue(manifest, swagger, operationId, definition);
   return {
     enabled: !!splitOnValue,
@@ -580,9 +584,10 @@ const isSplitOnSupported = (
   swagger?: SwaggerParser,
   operationId?: string,
   definition?: LogicAppsV2.OperationDefinition,
-  workflowKind?: WorkflowKind
+  workflowKind?: WorkflowKind,
+  forceEnableSplitOn?: boolean
 ): boolean => {
-  if (workflowKind === WorkflowKind.STATELESS) return false;
+  if (workflowKind === WorkflowKind.STATELESS && !forceEnableSplitOn) return false;
   const existingSplitOn = getSplitOn(manifest, swagger, operationId, definition);
   return isTrigger && (getSplitOnOptions(nodeOutputs, !!manifest).length > 0 || existingSplitOn.enabled);
 };