+++ libs/designer/src/lib/core/state/panel/panelInterfaces.ts
@@ -17,6 +17,7 @@ export interface PanelState {
   isLoading?: boolean;
   focusReturnElementId?: string;
   creatingConnection?: boolean;
+  selectedErrorsPanelTabId: string | undefined; // undefined will select first tab with > 0 messages
 }
 
 export interface RelationshipIds {