+++ libs/designer/src/lib/core/state/panel/panelSlice.ts
@@ -21,6 +21,7 @@ const initialState: PanelState = {
   creatingConnection: false,
   currentPanelMode: undefined,
   referencePanelMode: undefined,
+  selectedErrorsPanelTabId: undefined,
 };
 
 export const panelSlice = createSlice({
@@ -44,6 +45,7 @@ export const panelSlice = createSlice({
       state.addingTrigger = false;
       state.creatingConnection = false;
       state.selectedTabId = undefined;
+      state.selectedErrorsPanelTabId = undefined;
     },
     updatePanelLocation: (state, action: PayloadAction<PanelLocation | undefined>) => {
       if (action.payload && action.payload !== state.panelLocation) {
@@ -143,6 +145,16 @@ export const panelSlice = createSlice({
     setIsCreatingConnection: (state, action: PayloadAction<boolean>) => {
       state.creatingConnection = action.payload;
     },
+    selectErrorsPanelTab: (state, action: PayloadAction<string>) => {
+      state.selectedErrorsPanelTabId = action.payload;
+
+      LoggerService().log({
+        level: LogEntryLevel.Verbose,
+        area: 'Designer:Panel Slice',
+        message: action.type,
+        args: [action.payload],
+      });
+    },
   },
   extraReducers: (builder) => {
     builder.addCase(resetWorkflowState, () => initialState);
@@ -165,6 +177,7 @@ export const {
   selectPanelTab,
   setIsPanelLoading,
   setIsCreatingConnection,
+  selectErrorsPanelTab,
 } = panelSlice.actions;
 
 export default panelSlice.reducer;