+++ libs/designer/src/lib/ui/panel/nodeSearchPanel/nodeSearchPanel.tsx
@@ -1,3 +1,4 @@
+import { useHostOptions } from '../../../core/state/designerOptions/designerOptionsSelectors';
 import { useOperationVisuals } from '../../../core/state/operation/operationSelector';
 import { clearPanel } from '../../../core/state/panel/panelSlice';
 import { useNodeDisplayName, useNodeIds } from '../../../core/state/workflow/workflowSelectors';
@@ -43,12 +44,11 @@ const NodeSearchCard = ({ node, displayRuntimeInfo }: { node: string; displayRun
 };
 
 export type NodeSearchPanelProps = {
-  displayRuntimeInfo: boolean;
   focusReturnElementId: string | undefined;
 } & CommonPanelProps;
 
 export const NodeSearchPanel = (props: NodeSearchPanelProps) => {
-  const { displayRuntimeInfo } = props;
+  const { displayRuntimeInfo } = useHostOptions();
   const allNodeNames = useNodeIds();
   const [searchTerm, setSearchTerm] = useState<string | null>(null);
   const intl = useIntl();