+++ libs/designer/src/lib/ui/panel/panelRoot.tsx
@@ -1,14 +1,14 @@
 import type { AppDispatch } from '../../core';
-import { useHostOptions, useIsDarkMode } from '../../core/state/designerOptions/designerOptionsSelectors';
+import { useIsDarkMode } from '../../core/state/designerOptions/designerOptionsSelectors';
 import {
   useCurrentPanelMode,
-  useFocusReturnElementId,
   useIsLoadingPanel,
   useIsPanelCollapsed,
+  useFocusReturnElementId,
 } from '../../core/state/panel/panelSelectors';
 import { clearPanel } from '../../core/state/panel/panelSlice';
 import { ConnectionPanel } from './connectionsPanel/connectionsPanel';
-import { ErrorPanel } from './errorsPanel/errorsPanel';
+import { ErrorsPanel } from './errorsPanel/errorsPanel';
 import { NodeDetailsPanel } from './nodeDetailsPanel/nodeDetailsPanel';
 import { NodeSearchPanel } from './nodeSearchPanel/nodeSearchPanel';
 import { RecommendationPanelContext } from './recommendation/recommendationPanelContext';
@@ -35,7 +35,6 @@ const layerProps = {
 export const PanelRoot = (props: PanelRootProps): JSX.Element => {
   const { panelLocation, customPanelLocations } = props;
   const dispatch = useDispatch<AppDispatch>();
-  const { displayRuntimeInfo } = useHostOptions();
   const isDarkMode = useIsDarkMode();
 
   const collapsed = useIsPanelCollapsed();
@@ -105,13 +104,13 @@ export const PanelRoot = (props: PanelRootProps): JSX.Element => {
         ) : currentPanelMode === 'WorkflowParameters' ? (
           <WorkflowParametersPanel {...commonPanelProps} />
         ) : currentPanelMode === 'Discovery' ? (
-          <RecommendationPanelContext {...commonPanelProps} displayRuntimeInfo={displayRuntimeInfo} />
+          <RecommendationPanelContext {...commonPanelProps} />
         ) : currentPanelMode === 'NodeSearch' ? (
-          <NodeSearchPanel {...commonPanelProps} displayRuntimeInfo={displayRuntimeInfo} focusReturnElementId={focusReturnElementId} />
+          <NodeSearchPanel {...commonPanelProps} focusReturnElementId={focusReturnElementId} />
         ) : currentPanelMode === 'Connection' ? (
           <ConnectionPanel {...commonPanelProps} />
         ) : currentPanelMode === 'Error' ? (
-          <ErrorPanel {...commonPanelProps} />
+          <ErrorsPanel {...commonPanelProps} />
         ) : null // Caught above
       }
     </Panel>